requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-core==2.1.0
google-cloud-speech==2.11.1
google-cloud-texttospeech==2.7.1
google-cloud-translate==3.6.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.2
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SpeechRecognition==3.8.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
